some refactoring and renaming
