+++ modules/websocket_bridge/src/ipc/ipc_interface.cpp
@@ -1,3 +1,7 @@
+//=================================================================================================
+// Copyright (C) 2025 HEPHAESTUS Contributors
+//=================================================================================================
+
 #include "hephaestus/ipc/ipc_interface.h"
 
 #include <cstdio>
