+++ modules/websocket_bridge/src/websocket_bridge/bridge_config.cpp
@@ -1,3 +1,7 @@
+//=================================================================================================
+// Copyright (C) 2025 HEPHAESTUS Contributors
+//=================================================================================================
+
 #include "hephaestus/websocket_bridge/bridge_config.h"
 
 #include <rfl/yaml.hpp>
