+++ modules/websocket_bridge/src/websocket_bridge/bridge_state.cpp
@@ -1,5 +1,5 @@
 //=================================================================================================
-// Copyright (C) 2023-2024 HEPHAESTUS Contributors
+// Copyright (C) 2025 HEPHAESTUS Contributors
 //=================================================================================================
 
 #include "hephaestus/websocket_bridge/bridge_state.h"
