+++ modules/websocket_bridge/tests/CMakeLists.txt
@@ -23,3 +23,16 @@ define_module_test(
         $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/../include>
         PUBLIC_LINK_LIBS "")
 
+define_module_test(
+        NAME test_ipc_interface
+        SOURCES test_ipc_interface.cpp
+        PUBLIC_INCLUDE_PATHS
+        $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/../include>
+        PUBLIC_LINK_LIBS "")
+
+define_module_test(
+        NAME test_protobuf_serdes
+        SOURCES test_protobuf_serdes.cpp
+        PUBLIC_INCLUDE_PATHS
+        $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/../include>
+        PUBLIC_LINK_LIBS "")
