+++ modules/websocket_bridge/tests/test_bridge_config.cpp
@@ -1,3 +1,7 @@
+//=================================================================================================
+// Copyright (C) 2025 HEPHAESTUS Contributors
+//=================================================================================================
+
 #include <gtest/gtest.h>
 
 #include "hephaestus/websocket_bridge/bridge_config.h"
