+++ modules/websocket_bridge/tests/test_bridge_state.cpp
@@ -1,3 +1,7 @@
+//=================================================================================================
+// Copyright (C) 2025 HEPHAESTUS Contributors
+//=================================================================================================
+
 #include <optional>
 #include <string>
 #include <unordered_map>
