+++ modules/websocket_bridge/tests/test_ipc_interface.cpp
@@ -1,3 +1,7 @@
+//=================================================================================================
+// Copyright (C) 2025 HEPHAESTUS Contributors
+//=================================================================================================
+
 #include <chrono>
 #include <span>
 #include <thread>
